+++ package.json
@@ -14,6 +14,7 @@
     "@emotion/styled": "^11.14.1",
     "@fontsource/roboto": "^5.2.6",
     "@mui/icons-material": "^7.2.0",
+    "@mui/lab": "^7.0.1-beta.18",
     "@mui/material": "^7.2.0",
     "@mui/x-data-grid": "^8.9.2",
     "@toolpad/core": "^0.16.0",
@@ -23,7 +24,8 @@
     "react-cookie": "^8.0.1",
     "react-dom": "^17.0.0 || ^18.0.0 || ^19.0.0",
     "react-hook-form": "^7.61.1",
-    "react-router-dom": "^7.7.1"
+    "react-router-dom": "^7.7.1",
+    "usehooks-ts": "^3.1.1"
   },
   "peerDependencies": {
     "react": "^17.0.0 || ^18.0.0 || ^19.0.0",
