+++ src/components/financial_statement/index.tsx
@@ -1,16 +1,12 @@
 import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
 import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp';
-import Box from '@mui/material/Box';
-import Collapse from '@mui/material/Collapse';
+import { Typography } from '@mui/material';
 import IconButton from '@mui/material/IconButton';
-import Table from '@mui/material/Table';
-import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
-import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
-import Typography from '@mui/material/Typography';
 import { useState } from 'react';
 import type { formatType } from './CollapsibleTable';
+import DetailedRow from './DetailedRow';
 
 export function Row({ row }: { row: formatType }) {
   const [open, setOpen] = useState(false);
@@ -28,40 +24,13 @@ export function Row({ row }: { row: formatType }) {
           </IconButton>
         </TableCell>
         <TableCell component="th" scope="row">
-          {row.name}
+          <Typography sx={{ textTransform: 'capitalize' }} variant="h6">{row.name}</Typography>
         </TableCell>
         <TableCell align="right">{row.count}</TableCell>
         <TableCell align="right">{row.balance}</TableCell>
       </TableRow>
-      <TableRow>
-        <TableCell style={{ paddingBottom: 0, paddingTop: 0 }} colSpan={6}>
-          <Collapse in={open} timeout="auto" unmountOnExit>
-            <Box sx={{ margin: 1 }}>
-              <Typography variant="h6" gutterBottom component="div">
-                Breakdown
-              </Typography>
-              <Table size="small" aria-label="purchases">
-                <TableHead>
-                  <TableRow>
-                    <TableCell>Code of Account</TableCell>
-                    <TableCell align="right">Balance</TableCell>
-                  </TableRow>
-                </TableHead>
-                <TableBody>
-                  {[...row.coa_list].map(([coa, balance]) => (
-                    <TableRow key={coa}>
-                      <TableCell component="th" scope="row">
-                        {coa}
-                      </TableCell>
-                      <TableCell align="right">{balance}</TableCell>
-                    </TableRow>
-                  ))}
-                </TableBody>
-              </Table>
-            </Box>
-          </Collapse>
-        </TableCell>
-      </TableRow>
+
+      <DetailedRow row={row} open={open} />
     </>
   );
 }
