+++ src/components/ledger_form/LedgerItemsFormTable/index.tsx
@@ -10,29 +10,30 @@ import LedgerItemInputRow from './LedgerItemInputRow'
 export default function LedgerItemsFormTable() {    
 
     return (
-      <>
-        <TableContainer component={Paper}>
-          <Table sx={{ minWidth: 500 }} aria-label={`LedgerItemInputTable`}>
-            <TableBody>
-              <TableRow>
-                <TableCell sx={{ width: "10%" }} variant="head">
-                  {/* Blank cell for delete button */}
-                </TableCell>
-                <TableCell sx={{ width: "30%" }} variant="head">
-                  Code of Account
-                </TableCell>
-                <TableCell sx={{ width: "30%" }} variant="head">
-                  Amount
-                </TableCell>
-                <TableCell sx={{ width: "30%" }} variant="head">
-                  Balance Type
-                </TableCell>
-              </TableRow>
-              <LedgerItemInputRow />
-              <BalanceCheckRow />
-            </TableBody>
-          </Table>
-        </TableContainer>
-      </>
+      <TableContainer component={Paper}>
+        <Table sx={{ minWidth: 500 }} aria-label={`LedgerItemInputTable`}>
+          <TableBody>
+            <TableRow>
+              <TableCell sx={{ width: "10%" }} variant="head">
+                {/* Blank cell for delete button */}
+              </TableCell>
+              <TableCell sx={{ width: "30%" }} variant="head">
+                Code of Account
+              </TableCell>
+              <TableCell sx={{ width: "30%" }} variant="head">
+                Title
+              </TableCell>
+              <TableCell sx={{ width: "30%" }} variant="head">
+                Amount
+              </TableCell>
+              <TableCell sx={{ width: "30%" }} variant="head">
+                Balance Type
+              </TableCell>
+            </TableRow>
+            <LedgerItemInputRow />
+            <BalanceCheckRow />
+          </TableBody>
+        </Table>
+      </TableContainer>
     );
 }
