+++ src/components/ledger_list/FetchUtil.tsx
@@ -1,0 +1,52 @@
+import type { GridRowsProp } from "@mui/x-data-grid/models";
+import type { Dispatch, SetStateAction } from "react";
+import { axiosQueryClient } from "../../service/api";
+import type { LedgerResponse } from "./type";
+
+// define function to fetch data from the server
+export const fetchRows = async (
+    setRows: Dispatch<SetStateAction<GridRowsProp>>
+) => {
+    const data = await axiosQueryClient
+        .get<LedgerResponse[]>('/api/ledgers/all')
+        .then((res) => res?.data ?? null)
+        .catch(() => null);
+
+    if (data == null) {
+    console.warn(
+        "API returned null/invalid payload; rows cleared."
+    );
+    setRows([]);
+    return;
+    }
+
+    console.log('Before :' ,data);
+
+    // Flatten ledgerItems for each ledger into rows
+    const dataRows = data
+    .flatMap((ledger, idx) =>
+        ledger.ledgerItems.map((item, idy) => {
+        const isDebit = item.type == 'Debit';
+        const isCredit = item.type == 'Credit';
+
+        return {
+            id: `${idx}-${idy}`,
+            date: ledger.date,
+            coa: item.coa,
+            description: ledger.description,
+            debit: isDebit ? item.amount : 0,
+            credit: isCredit ? item.amount : 0,
+            balance: isDebit ? item.amount : -item.amount,
+            // calculate the balance of transaction
+            transaction_balance: ledger.ledgerItems.map((entry) => (
+                entry.amount / 2
+            )).reduce((sum, balance) => sum + balance, 0)
+            }
+        }
+        // sort ledgeritem by Code of Account
+        ).sort((a,b) => a.coa - b.coa)
+    );
+    setRows(dataRows);
+    
+    console.log("After :", dataRows);
+};
