+++ src/hooks/coa/context.ts
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import useProvideCoa from ".";
+
+export const CoaContext = createContext<ReturnType<typeof useProvideCoa> | null>(null);
