+++ src/hooks/coa/useCoa.ts
@@ -1,0 +1,7 @@
+import { useContext } from "react";
+import { CoaContext } from "./context";
+
+
+export const useCoa = () => {
+    return useContext(CoaContext);
+}
